backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pandas>=2.0.0
numpy>=1.24.0
geopy==2.4.1
gunicorn==21.2.0
scipy>=1.10.0
